virt-install: Use full path when auto-naming a storage pool
virt-install automatically creates a storage pool when you specify a
path to a disk that isn't already managed by a libvirt storage pool. The
current implementation just uses the basename of the parent directory
for the pool name. So `/path/to/parentdir/disk.img` will create a
storage pool named `parentdir`. If there is already a pool named
`parentdir`, we add a disambiguator to the name (e.g. `parentdir-1`).

Unfortunately, when a management app creates a bunch of guests at the
same time, this can result in some racy behavior. For example, if there
are multiple instances of virt-install running in parallel, we can end
up with an error such as that reported in the bug
https://bugzilla.redhat.com/show_bug.cgi?id=2177837:

ERROR Error: --disk path=/tmp/as64_47774/ISO/seed.iso,device=cdrom: Could not define storage pool: operation failed: pool 'ISO-973' already exists with uuid 42f1b413-4957-469c-b706-b60fcf8cd841

In this example, the management app apparently has created a directory
in /tmp for each guest being created (`/tmp/as64_47774`). This directory
contains a subdirectory named `ISO` that holds disk images. In the above
scenario, this was done enough times that the disambiguation number has
reached 973 (note the pool name `ISO-973`). This high number of name
collisions dramatically increases the chance of the race condition
described below:

One instance of `virt-install` notices that the specified disk image is
not located in a location managed by a libvirt pool so it tries to
automatically create a new storage pool. It tries to create a pool named
`ISO`, but notices that a pool of this name already exists, so it
tries `ISO-1`. It continues iterating until it finds that the first free
pool name of `ISO-973`. Unfortunately, another instance of
`virt-install` was also doing the same thing at the exact same time.
That instance also wants to create a pool named `ISO`, and it also finds
that the first free pool name is `ISO-973`. If the second `virt-install`
instance wins the race to creating the pool named `ISO-973`, the first
instance will get the error quoted above when it tries to define a new
pool with this name.

To reduce the chance of this error, we want to reduce the chance of
conflicts between pool names. By using the full path of the parent dir
as the name of the pool (with directory separators converted to
underscores), we nearly eliminate the chance of automatic pool name
conflicts. The only remaining possible name conflicts are:
 - a manually-created pool with the same name (though the likelihood
   that somebody would manually choose a name of this form is low)
 - directories that contain underscores in the same pattern as the
   directory structure (e.g. `/tmp/dir1/dir2/` and `/tmp/dir1_dir2`
   would result in the same automatically chosen name of
   `tmp_dir1_dir2`).

The risk of these potential name collisions is low enough that we
probably don't have to worry about them.

In the example above, we would have created a pool named
`tmp_as64_47774_ISO` rather than `ISO`

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=2177837

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>